requirements.txt
streamlit
tensorflow
numpy
opencv-python
Pillow
pytesseract